% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_eagles.R
\name{hello_eagles}
\alias{hello_eagles}
\title{Say hello to Eagle}
\usage{
hello_eagles(eagles, daytime = "auto")
}
\arguments{
\item{eagles}{character, the names of the EAGLEs to be greeted.}

\item{daytime}{character, optional, either "auto" to calculate daytime or "morning" or "evening", "afternoon"}
}
\value{
Nothing, it will just say hello.
}
\description{
This function says hello to Eagles depending on daytime.
}
