\name{make.knuth.random}
\alias{make.knuth.random}
\title{Make an LCG based on Knuth's heuristics.}
\usage{
  make.knuth.random(seed)
}
\arguments{
  \item{seed}{The initial seed}
}
\description{
  Make an LCG based on Knuth's heuristics.
}
\seealso{
  Other lcg: \code{\link{make.numerical.recipes.random}},
  \code{\link{make.pathological.random}},
  \code{\link{make.random}}
}

