\name{make.numerical.recipes.random}
\alias{make.numerical.recipes.random}
\title{Machen Sie einen LCG auf die Parameter in Numerical Recipes basiert.}
\usage{
  make.numerical.recipes.random(seed)
}
\arguments{
  \item{seed}{The initial seed}
}
\description{
  Machen Sie einen LCG auf die Parameter in Numerical
  Recipes basiert.
}
\seealso{
  Other lcg: \code{\link{make.knuth.random}},
  \code{\link{make.pathological.random}},
  \code{\link{make.random}}
}

