\name{make.pathological.random}
\alias{make.pathological.random}
\title{Machen Sie einen LCG mit einer pathologischen Periode der Länge 4.}
\usage{
  make.pathological.random(seed)
}
\arguments{
  \item{seed}{The initial seed}
}
\description{
  Machen Sie einen LCG mit einer pathologischen Periode der
  Länge 4.
}
\seealso{
  Other lcg: \code{\link{make.knuth.random}},
  \code{\link{make.numerical.recipes.random}},
  \code{\link{make.random}}
}

