\name{make.knuth.random}
\alias{make.knuth.random}
\title{Machen Sie einen LCG auf Knuth Heuristik basiert.}
\usage{
  make.knuth.random(seed)
}
\arguments{
  \item{seed}{The initial seed}
}
\description{
  Machen Sie einen LCG auf Knuth Heuristik basiert.
}
\seealso{
  Other lcg: \code{\link{make.numerical.recipes.random}},
  \code{\link{make.pathological.random}},
  \code{\link{make.random}}
}

