\name{make.knuth.random}
\alias{make.knuth.random}
\title{Make an LCG based on Knuth's heuristics.}
\usage{
  make.knuth.random()
}
\description{
  Make an LCG based on Knuth's heuristics.
}
\seealso{
  Other LCGs: \code{\link{make.numerical.recipes.random}},
  \code{\link{make.pathological.random}},
  \code{\link{make.random}}
}

