\name{make.numerical.recipes.random}
\alias{make.numerical.recipes.random}
\title{Make an LCG based on the parameters in Numerical Recipes.}
\usage{
  make.numerical.recipes.random()
}
\description{
  Make an LCG based on the parameters in Numerical Recipes.
}
\seealso{
  Other LCGs: \code{\link{make.knuth.random}},
  \code{\link{make.pathological.random}},
  \code{\link{make.random}}
}

