\name{make.pathological.random}
\alias{make.pathological.random}
\title{Make a random-number generator with a pathological period
(i.e. 4).}
\usage{
  make.pathological.random(seed)
}
\arguments{
  \item{seed}{The initial seed}
}
\description{
  Make a random-number generator with a pathological period
  (i.e. 4).
}

