\name{make.pathological.random}
\alias{make.pathological.random}
\title{Make an LCG with a pathological period of length 4.}
\usage{
  make.pathological.random(seed)
}
\arguments{
  \item{seed}{The initial seed}
}
\description{
  Make an LCG with a pathological period of length 4.
}
\seealso{
  Other lcg: \code{\link{make.knuth.random}},
  \code{\link{make.numerical.recipes.random}},
  \code{\link{make.random}}
}

