\name{make.pathological.random}
\alias{make.pathological.random}
\title{Make an LCG with a pathological period of length 4.}
\usage{
  make.pathological.random()
}
\description{
  Make an LCG with a pathological period of length 4.
}
\seealso{
  Other LCGs: \code{\link{make.knuth.random}},
  \code{\link{make.numerical.recipes.random}},
  \code{\link{make.random}}
}

